# Don't use lower() for citext columns